pyproject.toml
[project]
name = "nyx"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "advanced-alchemy>=1.8.0",
    "alembic>=1.17.1",
    "async-oauthlib>=0.0.9",
    "asyncpg>=0.30.0",
    "httpx==0.28.1",
    "litestar[standard]>=2.18.0",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "zitadel-client>=4.1.0b7",
]
